Promote `identity`, `record`, and `operation` modules
